src/api/views.py
from rest_framework import generics

from books.models import Book
from .permissions import IsAuthorOrReadOnly
from .serializers import BookSerializer

# Create your views here.
class BookAPIView(generics.ListAPIView):
    queryset = Book.objects.all()
    serializer_class = BookSerializer

class PostDetail(generics.RetrieveUpdateDestroyAPIView):
    permission_classes = (IsAuthorOrReadOnly,)
    queryset = Post.objects.all()
    serializer_class = PostSerializer