src/requirements.dev.txt
# General
pillow==7.0.0

# Django
Django==2.2.9

# Debug
django-debug-toolbar==2.2

# Secure Sensitive Constant
python-decouple==3.3

# Custom management commands
django-extensions==2.2.8

# Django Rest Framework
djangorestframework==3.11.0

# Django-cors-header
django-cors-headers==3.2.1

# Django REST Swagger
drf-yasg==1.17.0